backend/requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
httpx==0.28.1
redis==6.4.0
python-dotenv==1.1.1
pydantic==2.12.3
pydantic-settings==2.6.1
slowapi==0.1.9
clickhouse-driver==0.2.9
faker==37.11.0
pandas==2.3.3
numpy==2.3.4
groq>=0.6.0
sqlglot==27.28.1
pyyaml==6.0.2
